shsUpdate should take an unsigned int for length
Update shsUpdate to take an unsigned int length instead of signed.
The code already assumes that it is an integer >= 0. This change
cleans up a number of signed/unsigned warnings.

sha1/shs.h, shs.c, t_shs.c: change ints to unsigned int
dk/dk_encrypt.c: static function trunc_hmac change a signed int to unsigned
hash_provider/hash_sha1.c: Remove cast in argument to shsUpdate

ticket: new

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@17945 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/lib/crypto/hash_provider/hash_sha1.c
@@ -40,7 +40,7 @@ k5_sha1_hash(unsigned int icount, const krb5_data *input,
 
     shsInit(&ctx);
     for (i=0; i<icount; i++)
-	shsUpdate(&ctx, (unsigned char *) input[i].data, (int) input[i].length);
+	shsUpdate(&ctx, (unsigned char *) input[i].data, input[i].length);
     shsFinal(&ctx);
 
     for (i=0; i<(sizeof(ctx.digest)/sizeof(ctx.digest[0])); i++) {
